pyproject.toml
[project]
name = "documind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.16",
    "aiosignal==1.3.2",
    "alembic==1.15.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "authlib==1.3.1",
    "bcrypt==4.3.0",
    "blinker==1.9.0",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "cryptography==44.0.2",
    "dataclasses-json==0.6.7",
    "docx==0.2.4",
    "fastapi==0.115.12",
    "filelock==3.18.0",
    "flask==3.1.0",
    "flask-bcrypt==1.0.1",
    "flask-cors==5.0.1",
    "flask-jwt-extended==4.7.1",
    "flask-loginmanager==1.1.6",
    "flask-migrate==4.1.0",
    "flask-sqlalchemy==3.1.1",
    "frozenlist==1.6.0",
    "fsspec==2025.3.2",
    "greenlet==3.2.0",
    "grpcio==1.71.0",
    "grpcio-health-checking==1.71.0",
    "grpcio-tools==1.71.0",
    "gunicorn>=23.0.0",
    "h11==0.14.0",
    "httpcore==1.0.8",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.30.2",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "joblib==1.4.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.23",
    "langchain-community==0.3.21",
    "langchain-core==0.3.54",
    "langchain-ollama==0.3.2",
    "langchain-text-splitters==0.3.8",
    "langsmith==0.3.32",
    "lxml==5.3.2",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mpmath==1.3.0",
    "multidict==6.4.3",
    "mypy-extensions==1.0.0",
    "networkx==3.4.2",
    "numpy==2.2.4",
    "ollama==0.4.8",
    "orjson==3.10.16",
    "packaging==24.2",
    "pillow==11.2.1",
    "propcache==0.3.1",
    "protobuf==5.29.4",
    "psycopg2-binary==2.9.10",
    "pycparser==2.22",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pydantic-settings==2.8.1",
    "pyjwt==2.10.1",
    "pypdf2==3.0.1",
    "python-docx==1.1.2",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "safetensors==0.5.3",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "sentence-transformers==4.1.0",
    "setuptools==78.1.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "starlette==0.46.2",
    "sympy==1.13.1",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tokenizers==0.21.1",
    "torch==2.6.0",
    "tqdm==4.67.1",
    "transformers==4.51.3",
    "typing-extensions==4.13.2",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "validators==0.34.0",
    "weaviate==0.1.2",
    "weaviate-client==4.13.2",
    "werkzeug==3.1.3",
    "yarl==1.20.0",
    "zstandard==0.23.0",
]
